+++ coinbasepro/auth_client.py
@@ -60,7 +60,8 @@ class AuthenticatedClient(PublicClient):
     def get_account(self, account_id: str) -> Dict[str, Any]:
         """Get information for a single account.
 
-        Use this endpoint when you know the account_id.
+        Use this endpoint when you know the account_id. API key must
+        belong to the same profile as the account.
 
         Args:
             account_id: Account id for account you want to get.
@@ -177,8 +178,8 @@ class AuthenticatedClient(PublicClient):
         pending withdraw requests.
 
         As an order is filled, the hold amount is updated. If an order
-        is canceled, any remaining hold is removed. For a withdraw,
-        once it is completed, the hold is removed.
+        is canceled, any remaining hold is removed. For withdrawals,
+        the hold is removed after it is completed.
 
         The `type` field will indicate why the hold exists. The hold
         type is 'order' for holds related to open orders and 'transfer'
@@ -223,6 +224,154 @@ class AuthenticatedClient(PublicClient):
         )
         return (self._convert_dict(hold, field_conversions) for hold in r)
 
+    def get_account_transfers(
+        self, account_id: str, **kwargs
+    ) -> Iterator[Dict[str, Any]]:
+        """List past withdrawals and deposits for an account.
+
+        Args:
+            account_id: Account id to get holds of.
+            kwargs: Additional HTTP request parameters.
+
+        Yields:
+            Account deposits and withdrawals. Example::
+                [
+                    {
+                        "id": "19ac524d-8827-4246-a1b2-18dc5ca9472c",
+                        "type": "withdraw",
+                        "created_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                        "completed_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                        "canceled_at": None,
+                        "processed_at": None,
+                        "user_nonce": 1234,
+                        "amount": "1.00000000",
+                        "details": {
+                          "coinbase_account_id": "2b760113-fbba-5600-ac74-36482c130768",
+                          "coinbase_transaction_id": "5e697ed49f8417148f3366ea",
+                          "coinbase_payment_method_id": ""
+                        }
+                    },
+                    {
+                    ...
+                    }
+                ]
+
+        Raises:
+            See `get_products()`.
+
+        """
+        return self._get_transfers_helper(f"/accounts/{account_id}/transfers", **kwargs)
+
+    def get_all_transfers(self, **kwargs) -> Iterator[Dict[str, Any]]:
+        """List all  transfers.
+
+        Includes in-progress and completed transfers of funds in/out of
+        any of the user's accounts.
+
+        Args:
+            kwargs: Additional HTTP request parameters.
+
+        Yields:
+            Transfers. Example::
+                [
+                    {
+                        "id": "19ac524d-8827-4246-a1b2-18dc5ca9472c",
+                        "type": "withdraw",
+                        "created_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                        "completed_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                        "canceled_at": None,
+                        "processed_at": None,
+                        "user_nonce": 1234,
+                        "amount": "1.00000000",
+                        "details": {
+                          "coinbase_account_id": "2b760113-fbba-5600-ac74-36482c130768",
+                          "coinbase_transaction_id": "5e697ed49f8417148f3366ea",
+                          "coinbase_payment_method_id": ""
+                        }
+                    },
+                    {
+                    ...
+                    }
+                ]
+
+        Raises:
+            See `get_products()`.
+
+        """
+        return self._get_transfers_helper("/transfers", **kwargs)
+
+    def get_transfer(self, transfer_id: str) -> Dict[str, Any]:
+        """Get information on a single transfer.
+
+        Includes in-progress and completed transfers of funds in/out of
+        any of the user's accounts.
+
+        Args:
+            transfer_id: ID of the transfer to fetch.
+            kwargs: Additional HTTP request parameters.
+
+        Returns:
+            A transfer. Example::
+                {
+                    "id": "19ac524d-8827-4246-a1b2-18dc5ca9472c",
+                    "type": "withdraw",
+                    "created_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                    "completed_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                    "canceled_at": None,
+                    "processed_at": None,
+                    "user_nonce": 1234,
+                    "amount": "1.00000000",
+                    "details": {
+                      "coinbase_account_id": "2b760113-fbba-5600-ac74-36482c130768",
+                      "coinbase_transaction_id": "5e697ed49f8417148f3366ea",
+                      "coinbase_payment_method_id": ""
+                    }
+                }
+
+        Raises:
+            See `get_products()`.
+
+        """
+        field_conversions = {
+            "created_at": self._parse_datetime,
+            "completed_at": self._parse_datetime,
+            "canceled_at": self._parse_datetime,
+            "processed_at": self._parse_datetime,
+            "user_nonce": self._parse_optional_int,
+            "amount": Decimal,
+        }
+        r = self._send_message(
+            "get", f"/transfers/{transfer_id}", rate_limiter=self.a_rate_limiter
+        )
+        return self._convert_dict(r, field_conversions)
+
+    def get_address_book(self) -> List[Dict[str, Any]]:
+        """Get all addresses stored in the address book.
+
+        Returns:
+            List of addresses. Example::
+                [
+                    {
+                        "id": "e9c483b8-c502-11ec-9d64-0242ac120002",
+                        "address": "1JmYrFBLMSCLBwoL87gdQ5Qc9MLvb2egKk",
+                        "currency": "ETH",
+                        "label": "my wallet",
+                        "last_used": None,
+                        "address_book_added_at": datetime(2019, 3, 19, 22, 26, 22, 520000),
+                        "destination_tag": None
+                    },
+                    {
+                    ...
+                    }
+                ]
+        """
+        field_conversions = {
+            "last_used": self._parse_datetime,
+            "address_book_added_at": self._parse_datetime,
+        }
+        r = self._send_message("get", "/address-book", rate_limiter=self.a_rate_limiter)
+        return self._convert_list_of_dicts(r, field_conversions)
+
     def place_order(
         self,
         product_id: str,
@@ -232,7 +381,7 @@ class AuthenticatedClient(PublicClient):
         stop_price: Optional[Union[float, Decimal]] = None,
         client_oid: Optional[str] = None,
         stp: Optional[str] = None,
-        **kwargs
+        **kwargs,
     ) -> Dict[str, Any]:
         """Place an order.
 
@@ -584,7 +733,7 @@ class AuthenticatedClient(PublicClient):
         self,
         product_id: Optional[str] = None,
         status: Optional[Union[str, List[str]]] = None,
-        **kwargs
+        **kwargs,
     ) -> Iterator[Dict[str, Any]]:
         """List your current open orders.
 
@@ -821,6 +970,56 @@ class AuthenticatedClient(PublicClient):
         )
         return self._convert_dict(r, {"amount": Decimal})
 
+    def generate_crypto_address(self, account_id: str) -> Dict[str, Any]:
+        """Generate a one-time crypto address for depositing crypto.
+
+        Args:
+            account_id: ID of the coinbase account.
+
+        Returns:
+            The created address, plus metadata. Example::
+                {
+                    "id": "fc9fed1e-d25b-54d8-b52b-7fa250c9ae2d",
+                    "address": "0xd518A6B23D8bCA15B9cC46a00Be8a818E34Ca79E",
+                    "address_info": {
+                        "address": "0xd518A6B23D8bCA15B9cC46a00Be8a818E34Ca79E"
+                    },
+                    "name": "New exchange deposit address",
+                    "created_at": datetime(2019, 3, 19, 22, 26, 22),
+                    "updated_at": datetime(2019, 3, 19, 22, 26, 22),
+                    "network": "ethereum",
+                    "uri_scheme": "ethereum",
+                    "resource": "address",
+                    "resource_path": "/v2/accounts/faf4b657-a48c-56b2-bec2-77567e3f4f97/addresses/fc9fed1e-d25b-54d8-b52b-7fa250c9ae2d",
+                    "warnings": [
+                        {
+                            "title": "Only send Orchid (OXT) to this address",
+                            "details": "Sending any other digital asset, including Ethereum (ETH), will result in permanent loss.",
+                            "image_url": "https://dynamic-assets.coinbase.com/22b24ad9886150535671f158ccb0dd9d12089803728551c998e17e0f503484e9c38f3e8735354b5e622753684f040488b08d55b8ef5fef51592680f0c572bdfe/asset_icons/023010d790b9b1f47bc285802eafeab3d83c4de2029fe808d59935fbc54cdd7c.png"
+                        }
+                    ],
+                    "deposit_uri": "ethereum:0x4575f41308ec1483f3d399aa9a2826d74da13deb/transfer?address=0xd518A6B23D8bCA15B9cC46a00Be8a818E34Ca79E",
+                    "callback_url": null,
+                    "exchange_deposit_address": true
+                }
+
+        Raises:
+            See `get_products()`.
+
+        """
+        params = {"account_id": account_id}
+        field_conversions = {
+            "created_at": self._parse_datetime,
+            "updated_at": self._parse_datetime,
+        }
+        r = self._send_message(
+            "post",
+            f"/coinbase-accounts/{account_id}/addresses",
+            data=json.dumps(params),
+            rate_limiter=self.a_rate_limiter,
+        )
+        return self._convert_dict(r, field_conversions)
+
     def withdraw(
         self, amount: Union[float, Decimal], currency: str, payment_method_id: str
     ) -> Dict[str, Any]:
@@ -905,7 +1104,7 @@ class AuthenticatedClient(PublicClient):
 
     def withdraw_to_crypto(
         self, amount: Union[float, Decimal], currency: str, crypto_address: str
-    ):
+    ) -> Dict[str, Any]:
         """Withdraw funds to a crypto address.
 
         Args:
@@ -937,6 +1136,37 @@ class AuthenticatedClient(PublicClient):
             rate_limiter=self.a_rate_limiter,
         )
         return self._convert_dict(r, {"amount": Decimal})
+
+    def get_crypto_withdrawal_fee_estimate(
+        self, currency: str, crypto_address: str
+    ) -> Dict[str, Any]:
+        """Get the fee estimate for the crypto withdrawal to crypto address
+
+        Args:
+            currency: The type of currency (eg. 'BTC').
+            crypto_address: Crypto address to withdraw to.
+
+        Returns:
+            Fee estimate details. Example::
+                {
+                    "fee": Decimal(0.01)
+                }
+
+        Raises:
+            See `get_products()`.
+
+        """
+        params = {
+            "currency": currency,
+            "crypto_address": crypto_address,
+        }
+        r = self._send_message(
+            "get",
+            "/withdrawals/fee-estimate",
+            params=params,
+            rate_limiter=self.a_rate_limiter,
+        )
+        return self._convert_dict(r, {"fee": Decimal})
 
     def get_payment_methods(self) -> List[Dict[str, Any]]:
         """Get a list of your payment methods.
@@ -972,6 +1202,33 @@ class AuthenticatedClient(PublicClient):
             "get", "/coinbase-accounts", rate_limiter=self.a_rate_limiter
         )
         return self._convert_list_of_dicts(r, field_conversions)
+
+    def get_fees(self) -> Dict[str, Any]:
+        """Get current maker/taker fee rates and 30-day trailing volume.
+
+        Returns:
+            Fee details. Example::
+                {
+                    "maker_fee_rate": Decimal('0.0050'),
+                    "taker_fee_rate": Decimal('0.0050'),
+                    "usd_volume": Decimal('43806.92')
+                }
+
+        Raises:
+            See `get_products()`.
+
+        """
+        field_conversions = {
+            "maker_fee_rate": Decimal,
+            "taker_fee_rate": Decimal,
+            "usd_volume": Decimal,
+        }
+        r = self._send_message(
+            "get",
+            "/fees",
+            rate_limiter=self.a_rate_limiter,
+        )
+        return self._convert_dict(r, field_conversions)
 
     def create_report(
         self,
@@ -1091,7 +1348,9 @@ class AuthenticatedClient(PublicClient):
         )
         return self._convert_dict(field_conversions, r)
 
-    def _get_account_helper(self, account_id):
+    def _get_account_helper(
+        self, account_id: str
+    ) -> Union[Dict[str, Any], List[Dict[str, Any]]]:
         field_conversions = {"balance": Decimal, "available": Decimal, "hold": Decimal}
         r = self._send_message(
             "get", "/accounts/" + account_id, rate_limiter=self.a_rate_limiter
@@ -1101,3 +1360,19 @@ class AuthenticatedClient(PublicClient):
             return self._convert_list_of_dicts(r, field_conversions)
         else:
             return self._convert_dict(r, field_conversions)
+
+    def _get_transfers_helper(
+        self, endpoint: str, **kwargs
+    ) -> Iterator[Dict[str, Any]]:
+        field_conversions = {
+            "created_at": self._parse_datetime,
+            "completed_at": self._parse_datetime,
+            "canceled_at": self._parse_datetime,
+            "processed_at": self._parse_datetime,
+            "user_nonce": self._parse_optional_int,
+            "amount": Decimal,
+        }
+        r = self._send_paginated_message(
+            endpoint, params=kwargs, rate_limiter=self.a_rate_limiter
+        )
+        return (self._convert_dict(hold, field_conversions) for hold in r)